changed to reflect SD and pass tests
This version is NOT final. JavaDoc. I will be collaborating with David
to see if any changes are needed and same with the version he made.
Also, I David and I will be collaborating with the test team to make
sure the necessary parts are tested and all the tests pass. This should
be finalized by Monday night 8/10 with the Javadoc added too.